Move StateEffectDispatch and "use" function to UnidirectionalViewModel.kt for reference comment.
